feat: "Add support for compound operators and inline comments"
This commit introduces support for compound operators (!=, ==, <=, >=) and single-line comments (//). Case statement in the scanner now has additional cases handling '!','=' ,'<', '>', and '/'. For '/', apart from being recognized as a FORWARD_SLASH, it's also checked for the following slash indicating an inline comment. In that case, code is added to advance the scanner to the end of the line.

New private methods 'match' and 'peek' were created in the process to clean up and increase the readability of the code. Match method checks if the next character is the one that's been expected and if it's so, it advances scanner position and returns true. 'Peek' method predicts the next character without advancing the scanner.